+++ DeBroglie.Console/Export/IExporter.cs
@@ -1,5 +1,4 @@
 ﻿using DeBroglie.Console.Config;
-using DeBroglie.Console.Export;
 using DeBroglie.Models;
 
 namespace DeBroglie.Console.Export
@@ -7,6 +6,5 @@ namespace DeBroglie.Console.Export
     public interface IExporter
     {
         void Export(TileModel model, TilePropagator tilePropagator, string filename, DeBroglieConfig config, ExportOptions exportOptions);
-
     }
 }